ui improvement

+++ app/src/common/parser.h
@@ -33,7 +33,7 @@ parser_error_t parser_parse(parser_context_t *ctx,
 parser_error_t parser_validate(const parser_context_t *ctx);
 
 //// returns the number of items in the current parsing context
-parser_error_t parser_getNumItems(const parser_context_t *ctx, uint16_t *num_items);
+parser_error_t parser_getNumItems(const parser_context_t *ctx, uint8_t *num_items);
 
 // retrieves a readable output for each field / page
 parser_error_t parser_getItem(const parser_context_t *ctx,
