+++ app/src/common/tx.c
@@ -93,7 +93,7 @@ const char *tx_parse() {
     return NULL;
 }
 
-tx_error_t tx_getNumItems(uint16_t *num_items) {
+tx_error_t tx_getNumItems(uint8_t *num_items) {
     parser_error_t err = parser_getNumItems(&ctx_parsed_tx, num_items);
 
     if (err != parser_ok) {
@@ -109,7 +109,7 @@ tx_error_t tx_getItem(int8_t displayIdx,
                       uint8_t pageIdx, uint8_t *pageCount) {
     tx_error_t err = tx_no_error;
 
-    uint16_t numItems = 0;
+    uint8_t numItems = 0;
     err = tx_getNumItems(&numItems);
     if (err != tx_no_error) {
         return err;
