+++ tests/utils/common.cpp
@@ -26,7 +26,7 @@
 std::vector<std::string> dumpUI(parser_context_t *ctx,
                                 uint16_t maxKeyLen,
                                 uint16_t maxValueLen) {
-    uint16_t numItems;
+    uint8_t numItems;
 
     parser_getNumItems(ctx, &numItems);
